Fixing a bug in sorting the y values in the same order as x
Previous code tried to be clever, but failed to keep track of the fact that if we sort x, is.unsorted(x) will start to evaluate to false later. 

IMO, the flow is better working with 1 variable at a time, so I added a flag to remember the sorted status of the data.